requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
requests==2.32.3
setuptools==78.1.0
sqlparse==0.5.3
urllib3==2.4.0
